requirements.txt
asgiref==3.7.2
Django==5.0
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-vision==3.5.0
googleapis-common-protos==1.62.0
django-modeltranslation==0.18.11
googletrans==4.0.0-rc1
